requirements.txt
# Do not include azure-functions-worker in this file
# The Python Worker is managed by the Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
azure-cosmos==4.3.1
azure-storage-blob==12.17.0
openai
httpx==0.27.2
arxiv==1.4.7
requests==2.31.0
beautifulsoup4==4.12.2